ANDROID: cpufreq: concurrent_policy_time by pid
In order to model the energy used by the cpu,
we need to expose cpu time information to
userspace. Output the time each pid spent running
on group of cpu's with the same policy with 1
to num_related_cpus online and not idle.

We can use this information to blame pids and
uids for the energy they are using.

For instance, if 2 cores that are a part of the
same policy are online and active for 50ms and
the pid is running on one of the cores, the pid
should be blamed for 1/2 of the energy that is
used everytime one or more cpus in the same policy
is active. Using other proc files, the pid can
also be blamed for the cost of keeping the entire
cpu unsuspended and the cost of the individual
cores it is using.

b/68399339

Test: cat /proc/<PID>/concurrent_policy_time

Change-Id: I1702c59cb2728a18a0ea67dae7c9a0565eee06ff
Signed-off-by: Marissa Wall <marissaw@google.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>