qcacld-2.0: Do not force-enable WLAN_THERMAL_SHUTDOWN
This caused FEATURE_WLAN_THERMAL_SHUTDOWN to be defined in Kbuild, which
leads to the dreaded 5-second wlan_auto_shutdown_wl wakelock.
The wlan_auto_shutdown_wl wakelock seems to be used primarily by mobile
routers instead according to my interpretation of Kbuild, so I believe
it is fine to not have this wakelock. It is also unlikely that phones
will frequently heat up to the point where WLAN needs to be shut down,
compared to mobile routers.

Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>