qcom-cpufreq: add underclock code
Apparently, the higher stock frequencies on both MSM8996 and MSM8996pro
are
unstable, even with a kernel binary that the OEM created. Running AnTuTu
with thermal throttling disabled (to force the CPUs to remain at their
respective maxfreqs) _always_ crashes MSM8996pro at its stock maxfreqs,
and
occasionally crashes MSM8996 at its stock maxfreqs. A workaround to this
is
to underclock the power cluster (pwrcl) on both chipsets.

On MSM8996, the perfcl's stock maxfreq of 2150 MHz is stable when the
pwrcl
is underclocked to 1478 MHz. Therefore, allow the underclock to be
disabled
_only_ for the perfcl on MSM8996 (pwrcl must always remain
underclocked).

This is a very hacky solution to a very serious bug, so add TODOs to
investigate this further. Kernel logs are clean for these crashes, and
they could still be reproduced with a 20W wall charger connected to the
device, so they are not caused by a battery issue.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>